+++ src/plugins/nonebot_plugin_github/help.yaml
@@ -1,7 +1,7 @@
 plugin: github
 commands:
-  pacman:
+  github:
     description: help.description
     details: help.details
     usages:
-      - help.usage
+      - help.usage1
