+++ src/plugins/nonebot_plugin_github/lang.py
@@ -1,0 +1,2 @@
+from nonebot_plugin_larklang import LangHelper
+lang = LangHelper()
