+++ src/templates/github.html.jinja
@@ -1,0 +1,117 @@
+{% extends base %}
+{% block body %}
+{% if data.type == "repo" %}
+<div class="row">
+    <div class="col-2">
+        <img src="{{ data.avatar }}" class="rounded-circle" height="100%">
+    </div>
+    <div class="col-10">
+        <h2>{{ data.repo }}</h2>
+        {{ data.description }}
+    </div>
+</div>
+<div class="row">
+    <div class="col">
+        <h3>{{ text.repo.star }}</h3>
+        <p>{{ data.stars }}</p>
+        <h3>{{ text.repo.forks }}</h3>
+        <p>{{ data.forks }}</p>
+    </div>
+    <div class="col">
+        <h3>{{ text.repo.issues }}</h3>
+        <p>{{ data.issues }}</p>
+        <h3>{{ text.repo.language }}</h3>
+        <p>{{ data.language }}</p>
+    </div>
+</div>
+
+{% else if data.type == "issue" %}
+<h2>{{ data.title }}</h2>
+{{ data.repo }}<strong>#{{ data.number }}</strong>
+<hr>
+<div class="row">
+    <div class="col-2">
+        <img src="{{ data.avatar }}" class="rounded-circle" width="100%">
+    </div>
+    <div class="col-10">
+        <h4>{{ data.user }}</h4>
+        <pre>{{ data.body }}</pre>
+    </div>
+</div>
+<hr>
+<ul>
+    <li>
+        {{ text.issues.state }}
+        {% if data.state == "open" %}
+            <span class="badge bg-success">{{ text.issues.open }}</span>
+        {% else %}
+            <span class="badge bg-secondary">{{ text.issues.closed }}</span>
+        {% endif %}
+    </li>
+    <li>
+        {{ text.issues.labels }}
+        <ul>
+            {% for label in data.labels %}
+            <li>{{ label }}</li>
+            {% endfor %}
+        </ul>
+    </li>
+    <li>
+        {{ text.issues.updated_at }}
+        {{ data.updated_at}}
+    </li>
+</ul>
+
+
+{% else if data.type == "discussion" %}
+<h2>{{ data.title }}</h2>
+{{ data.repo }}<strong>#{{ data.number }}</strong>
+<hr>
+<div class="row">
+    <div class="col-2">
+        <img src="{{ data.avatar }}" class="rounded-circle" width="100%">
+    </div>
+    <div class="col-10">
+        <h4>{{ data.user }}</h4>
+        <small>{{ data.updated_at }} at {{ data.category }}</small>
+        <br>
+        <pre>{{ data.body }}</pre>
+    </div>
+</div>
+<hr>
+<ul>
+    <li>
+        {{ text.issues.state }}
+        {% if data.state == "open" %}
+            <span class="badge bg-success">{{ text.issues.open }}</span>
+        {% else %}
+            <span class="badge bg-secondary">{{ text.issues.closed }}</span>
+        {% endif %}
+    </li>
+    <li>
+        {{ text.issues.labels }}
+        <ul>
+            {% for label in data.labels %}
+            <li>{{ label }}</li>
+            {% endfor %}
+        </ul>
+    </li>
+</ul>
+
+{% else if data.type == "pull" %}
+<h2>{{ data.title }}</h2>
+{{ data.repo }}<strong>#{{ data.number }}</strong>
+
+<div class="row">
+    <div class="col-2">
+        <img src="{{ data.avatar }}" class="rounded-circle" width="100%">
+    </div>
+    <div class="col-10">
+        <p><span class="h4">{{ data.user }}</span> {{ text.pull.want_to_merge}} {{data.from}} {{text.pull.into}} {{ data.to }}</p>
+        <pre>{{ data.body }}</pre>
+    </div>
+</div>
+{{text.pull.updated_at}}{{data.updated_at}}
+
+{% endif %}
+{% endblock body %}
